requirements.txt
yandex-music
python-dotenv
npyscreen